requirements.txt
asgiref==3.2.7
autopep8==1.5.2
braintree==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-js-asset==1.2.2
django-multi-email-field==0.6.1
django-payments==0.13.0
django-payments-razorpay==1.0
django-tinymce4-lite==1.8.0
idna==2.9
jsmin==2.2.2
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
suds-jurko==0.6
urllib3==1.25.9
whitenoise==5.1.0
xmltodict==0.12.0